refactor(provider): move provider to internal namespace
