+++ internal/provider/provider.go
@@ -1,4 +1,4 @@
-package ohdear
+package provider
 
 import (
 	"context"
@@ -30,36 +30,32 @@ func init() {
 
 func New(version string) func() *schema.Provider {
 	return func() *schema.Provider {
-		return Provider()
-	}
-}
-
-func Provider() *schema.Provider {
-	return &schema.Provider{
-		Schema: map[string]*schema.Schema{
-			"api_token": {
-				Type:        schema.TypeString,
-				Required:    true,
-				Description: "Oh Dear API token. If not set, uses `OHDEAR_TOKEN` env var.",
-				DefaultFunc: schema.EnvDefaultFunc("OHDEAR_TOKEN", nil),
+		return &schema.Provider{
+			Schema: map[string]*schema.Schema{
+				"api_token": {
+					Type:        schema.TypeString,
+					Required:    true,
+					Description: "Oh Dear API token. If not set, uses `OHDEAR_TOKEN` env var.",
+					DefaultFunc: schema.EnvDefaultFunc("OHDEAR_TOKEN", nil),
+				},
+				"api_url": {
+					Type:        schema.TypeString,
+					Optional:    true,
+					Description: "Oh Dear API URL. If not set, uses `OHDEAR_API_URL` env var. Defaults to `https://ohdear.app`.",
+					DefaultFunc: schema.EnvDefaultFunc("OHDEAR_API_URL", "https://ohdear.app"),
+				},
+				"team_id": {
+					Type:        schema.TypeInt,
+					Optional:    true,
+					Description: "The default team ID to use for sites. If not set, uses `OHDEAR_TEAM_ID` env var.",
+					DefaultFunc: schema.EnvDefaultFunc("OHDEAR_TEAM_ID", 0),
+				},
 			},
-			"api_url": {
-				Type:        schema.TypeString,
-				Optional:    true,
-				Description: "Oh Dear API URL. If not set, uses `OHDEAR_API_URL` env var. Defaults to `https://ohdear.app`.",
-				DefaultFunc: schema.EnvDefaultFunc("OHDEAR_API_URL", "https://ohdear.app"),
+			ResourcesMap: map[string]*schema.Resource{
+				"ohdear_site": resourceOhdearSite(),
 			},
-			"team_id": {
-				Type:        schema.TypeInt,
-				Optional:    true,
-				Description: "The default team ID to use for sites. If not set, uses `OHDEAR_TEAM_ID` env var.",
-				DefaultFunc: schema.EnvDefaultFunc("OHDEAR_TEAM_ID", 0),
-			},
-		},
-		ResourcesMap: map[string]*schema.Resource{
-			"ohdear_site": resourceOhdearSite(),
-		},
-		ConfigureContextFunc: providerConfigure,
+			ConfigureContextFunc: providerConfigure,
+		}
 	}
 }
