+++ internal/provider/provider_test.go
@@ -1,4 +1,4 @@
-package ohdear
+package provider
 
 import (
 	"context"
