+++ internal/provider/resource_site.go
@@ -1,4 +1,4 @@
-package ohdear
+package provider
 
 import (
 	"context"
