+++ internal/provider/resource_site_test.go
@@ -1,4 +1,4 @@
-package ohdear
+package provider
 
 import (
 	"fmt"
